+++ src/pub/proxies.js
@@ -124,6 +124,27 @@ class Type_cell extends React.Component {
     }
 }
 
+class Browser_cell extends Pure_component {
+    open_browser = e=>{
+        e.stopPropagation();
+        const _this = this;
+        this.etask(function*(){
+            const url = `/api/browser/${_this.props.proxy.port}`;
+            const res = yield window.fetch(url);
+            if (res.status==206)
+                $('#fetching_chrome_modal').modal();
+        });
+    };
+    render(){
+        const class_names = 'btn btn_lpm btn_lpm_small';
+        const tooltip = 'Open browser configured with this port';
+        return is_local() && <T>{t=><Tooltip title={t(tooltip)}>
+              <button className={class_names}
+                onClick={this.open_browser}>Browser</button>
+            </Tooltip>}</T>;
+    }
+}
+
 class Port_cell extends Pure_component {
     state = {expanded: this.props.proxy.expanded};
     toggle = e=>{
@@ -249,7 +270,6 @@ const columns = [
         render: Type_cell,
         tooltip: 'Type of connected proxy - Luminati proxy or external proxy '
             +'(non-Luminati)',
-        default: true,
         ext: true,
         width: 70,
     },
@@ -433,8 +453,19 @@ const columns = [
         ext: true,
         tooltip: 'Number of all requests sent from this proxy port',
         dynamic: true,
-        grow: 0,
+        grow: 1,
         width: 60,
+    },
+    {
+        key: 'browser',
+        title: 'Browser',
+        render: Browser_cell,
+        default: true,
+        tooltip: 'Open browser configured with this port',
+        dynamic: true,
+        width: 80,
+        grow: 0,
+        hide_col_title: true,
     },
 ];
 
@@ -814,7 +845,7 @@ const Proxies = withRouter(class Proxies extends Pure_component {
                         {cols.map(col=>
                           <Column key={col.key}
                             cellRenderer={this.cell_renderer.bind(this)}
-                            label={t(col.title)}
+                            label={col.hide_col_title ? '' : t(col.title)}
                             className="cp_td"
                             dataKey={col.key}
                             flexGrow={col.grow!==undefined ? col.grow : 1}
@@ -1017,16 +1048,6 @@ class Actions extends Pure_component {
             yield _this.props.update_proxies();
         });
     };
-    open_browser = e=>{
-        e.stopPropagation();
-        const _this = this;
-        this.etask(function*(){
-            const url = `/api/browser/${_this.props.proxy.port}`;
-            const res = yield window.fetch(url);
-            if (res.status==206)
-                $('#fetching_chrome_modal').modal();
-        });
-    };
     open_delete_dialog_with_port = e=>{
         e.stopPropagation();
         this.props.open_delete_dialog([this.props.proxy]);
@@ -1048,11 +1069,6 @@ class Actions extends Pure_component {
             <Action_icon id="refresh" scrolling={this.props.scrolling}
               on_click={this.refresh_sessions}
               tooltip="Refresh Sessions"/>
-            {is_local() &&
-              <Action_icon id="browser" scrolling={this.props.scrolling}
-                on_click={this.open_browser}
-                tooltip="Open browser configured with this port"/>
-            }
           </div>;
     }
 }